+++ index.ts
@@ -3373,6 +3373,26 @@ namespace Enumerable {
     }
 
     /**
+     * Returns a sequence of random numbers.
+     * 
+     * @param {number} [count] The maximum number of items.
+     *                         If not defined, the sequence will become infinitely.
+     * @param {(randomValue: number, index: number) => number} [valueProvider] A custom function for providing a random number.
+     * 
+     * @return {IEnumerable<number>} The sequence of random numbers.
+     */
+    export function random(count?: number,
+                           valueProvider?: (randomValue: number, index: number) => number): IEnumerable<number> {
+        if (!valueProvider) {
+            valueProvider = (randVal) => randVal;
+        }
+        
+        return build((cancel, index) => valueProvider(Math.random(),
+                                                      index), 
+                     count);
+    }
+
+    /**
      * Creates a range of numbers.
      * 
      * @param {number} start The start value.